global: preparation for beta release
* Moves `cli.py` to `__main__.py` and `cmd.py` to `cli.py` to follow
  naming conventions.

* Adds PyPy to testing matrix on Travis CI.
