+++ docs/api.rst
@@ -6,11 +6,9 @@ Application and CLI factory
 
 .. automodule:: invenio_base.app
    :members:
-   :undoc-members:
 
 WSGI factory
 ------------
 
 .. automodule:: invenio_base.wsgi
    :members:
-   :undoc-members:
